Create informaton.yaml
I am developing a project based on the ton blockchain. I ask you to approve my coin so that it is considered confirmed in tonkipper and similar wallets. It is not a currency that can be used to buy anything. If you want to contact me, you can write to the email address listed on my website to participate in the project.